+++ packages/moonstone/Marquee/MarqueeDecorator.js
@@ -319,6 +319,9 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			if (node && this.distance == null && !this.props.disabled && !this.props.marqueeDisabled) {
 				this.distance = this.calculateDistance(node);
 				this.contentFits = !this.shouldAnimate(this.distance);
+				if(this.sync && this.contentFits) {
+					this.context.complete(this);
+				}
 				this.setState({
 					overflow: this.calculateTextOverflow(this.distance)
 				});
@@ -367,7 +370,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		 * @returns	{undefined}
 		 */
 		start = (delay = this.props.marqueeDelay) => {
-			if (this.contentFits) {
+			if (this.props.disabled || this.props.marqueeDisabled || this.contentFits) {
 				// if marquee isn't necessary (contentFits), do not set `animating` but return
 				// `true` to mark it complete if its synchronized so it doesn't block other
 				// instances.